Move files in Google MC Account.
