+++ js/src/setup-mc/setup-stepper/setup-accounts/google-mc-account/create-account.js
@@ -2,9 +2,9 @@
  * Internal dependencies
  */
 import CreateAccountCard from './create-account-card';
-import CreatingCard from '../creating-card';
-import ReclaimUrlCard from '../reclaim-url-card';
-import useCreateMCAccount from '../useCreateMCAccount';
+import CreatingCard from './creating-card';
+import ReclaimUrlCard from './reclaim-url-card';
+import useCreateMCAccount from './useCreateMCAccount';
 
 /**
  * Create Account flow.
